[10/16] DirAsDSK: Removed truncateCorrespondingFile()
Before this patch there were different code paths in writeDIREntry() to handle
a growing or a shrinking file. This is (no longer) needed, this patch treats
both in the same way.

(I didn't spend too much time investigating this, so this paragraph may not be
100% correct). Historical note: the first patch in this series in
extractCacheToFile() changed the file-open-mode from 'CREATE' (create if does
not exist yet) to 'TRUNCATE' (always start from a zero-length file). Without
this change the different code path may indeed have been needed. Though it was
still buggy because the code checked for changes in msx filesize to decide how
to growm/shrink the host file, while it might have been possible that the host
filesize was already changed without the msx side knowing about that change.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12851 7512d703-071c-0410-b5c2-d2bdd895b7d1