[13/16] DirAsDSK: Don't have separate buffers for boot, fat, dir
After the previous patches in this series, we now have permanent buffer space
for all the data sectors in the virtual disk. And the data and meta-data of the
virtual disk is now always updated atomically (I mean on a sync host->msx, of
course the msx must still write the sectors separately).

Before this patch we had separate storage for the boot sector, the 2 FATs, the
directory entries and the data sectors. There's no longer a reason for this, so
this patch unifies all this storage in a single virtual disk image. The format
of this information has now exactly the same layout as on a FAT12 filesystem
(was not yet 100% the case for the directory entries).

This unification allows for some nice cleanups in the code.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12854 7512d703-071c-0410-b5c2-d2bdd895b7d1