[15/16] DirAsDSK: Get rid of sectorMap
We (used to) have a data structure (the 'reverse-sector-map') that for each
data sector in the virtual disk images indicates to which directory entry and
at which offset in the corresponding file this sector is mapped. This structure
was used when the msx writes to a data sector, we can then quickly lookup to
which host file (and at what offset) this data chunk must be written.

This patch removes that data structure. We actually still need this
information, but it is possible to recalculate it from the current status of
the FAT and DIR sectors, so that's what this patch does. This is less efficient
(but still fast enough), but the main advantage is that we don't need any code
anymore to keep this data structure up-to-date (because this touches quite a
few places).

Actually when a FAT sector is modified we also need to know which cluster
(=position in the FAT) belongs to which file. We could have used the same
reverse sector map data structure for this problem, but instead the existing
code already calculated this from the FAT and DIR sectors. So the routines to
extract this information already existed, we now reuse them for data sector
writes. (And extract this functionality in actual subroutines, before it was
integrated in the write-fat routines).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12856 7512d703-071c-0410-b5c2-d2bdd895b7d1