[16/16] DirAsDSK: Don't treat FAT2 special
In patch 9 in this series we already tweaked the FAT-sync behavior, the
algorithm was to buffer changes on writes to FAT1 and only execute these
changes on (any) write to FAT2. The idea behind this algorithm was to avoid
doing updates for partially updated FATs (so 1 FAT sector already written, the
other 2 FAT sectors not yet written).

Though this doesn't work. I mean the DirAsDSK code does functionally work, but
the detection mechanism for partial FAT writes doesn't actually work. This is
because of the track level emulation (see patch 1 in this series). In case the
msx writes to sector 1 (= 1st FAT sector), the DirAsDSK back-end sees 9 sector
reads followed by 9 sector writes (= write to all sectors of FAT1 and FAT2). So
on each sector write we would detect a 'finished FAT update` condition. The
fact that DirAsDSK still works means that we don't actually need this
detection.

So this patch removes the detection mechanism. Read and writes to FAT2 are now
not treated specially anymore. DirAsDSK now simply memcpy's read/writes to FAT2
from/to the virtual disk image, but it doesn't try to interpret the data in
these sectors in any way. It also doesn't trigger any sync actions anymore on
writes to FAT2.

We are allowed to make this simplification for a similar reason as why we were
allowed to simplify writes to the DIR sectors (patch 11). Sometimes the msx
needs to write multiple FAT sectors to go from one consistent state to the next
(just like in patch 11 the msx needed to write multiple FAT or DIR sectors). We
now sync msx->host on all intermediate steps. So that means that for (short)
periods, files in the host may be truncated (because of an incomplete FAT
chain), but once the msx action is finished, the host filesystem should again
exactly reflect the state of the msx filesystem.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12857 7512d703-071c-0410-b5c2-d2bdd895b7d1