Fixed(?) DirAsDSK::SECTOR_SIZE related link error (vc++)
There was the following comment in the code:
    // REDEFINE (first definition is in SectorAccessibleDisk.hh)
    //  to solve link error with i686-apple-darwin9-gcc-4.0.1
And the constant SECTOR_SIZE was thus both defined in the base class
SectorAccessibleDisk and redefined in the subclass DirAsDSK. It's possible
gcc-4.0.1 really had a bug that required this workaround, but I now think it's
more likely we didn't properly define the constant in the base class.

When in a header file you have something like:
    class MyClass {
        static const int MY_CONSTANT = 42;
    };
you actually should in the .cc file put something like:
    const int MyClass::MY_CONSTANT;
Otherwise you could get link errors. The annoying thing is that in the vast
majority of the cases the code compiles/links fine without that extra line in
the .cc file. It depends on the compiler version and optimization settings.

Now back to this patch. I'm not exactly sure why vc++ gave this link error (I
think it should be OK, to have a constant with the same name in both the base
and the sub class). But the problem should go away if we only have one location
where the SECTOR_SIZE constant is defined. It's anyway how the code was meant
to be written.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12858 7512d703-071c-0410-b5c2-d2bdd895b7d1