Fix vc++ link error for real
In revision 12858 I tried to fix a (vc++ specific) link error. Though the error
was still present after that patch. Now I investigated it further with Vampier
and this patch should fix the problem for real. (It turned out to be caused by
non-standard behavior of vc++, see comments in code for details).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12860 7512d703-071c-0410-b5c2-d2bdd895b7d1