Fixed runtime warnings produced by DirAsDsk
In some scenarios DirAsDsk would give warnings like these (the exact message is
slightly different in windows/linux):
  warning: Couldn't write to file /home/wouter/dirasdsk/: Error reading file
(With '/home/wouter/dirasdsk' being the host directory used as dirasdsk). Apart
from this warning everything did work correctly. I could trigger this warning
with the following scenario:
- Start openMSX using a turboR msx machine and insert an empty host directory
  as DirAsDsk
- execute SAVE"FOO.BAS"   <-- save a (empty) BASIC prog
- execute KILL"FOO.BAS"   <-- delete file again (disk contains a deleted file)
- execute SAVE"BAR.BAS"   <-- create (any) new file

The warning was triggered when a data sector was written that used to belong to
a deleted MSX file: the directory entry starts with a 0xE5 character (indicates
a deleted file), but the start cluster still refers to the just written data
sector. The bug was that we forgot to check whether this directory entry
actually has an associated host file.  That's not the case in this scenario
(the associated host filename (correctly) is an empty string), so we tried to
open the dirasdsk host directory itself as a file and this triggered the
warning. Fixed by adding this check.

The problem triggered (only?) on turboR because the turboR disk rom first write
to the data sectors before updating the directory entry of the new file. The
disk rom of a Philips NMS 8250 first writes the directory entry with some fields
(e.g. the start cluster) still 0, followed by writes to the data sector(s) and
last it rewrites the directory entry with all fields filled in.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12873 7512d703-071c-0410-b5c2-d2bdd895b7d1