Don't try to uncompress files in DirAsDsk
The file handling routines in openMSX (by default) transparently handle (g)zip
compressed files. Usually this is what you want, though not when you read/write
host files from a DirAsDsk. In that case the host and the msx files should
always be the same (the msx file should not be an un(g)zipped version of the
host file). There was one location in the DirAsDsk code where we forgot to
disable the automatic un(g)zip mechanism.

I discovered this problem while working on the previous commit (warning when
writing to dirasdsk in some scenarios). Although this warning triggered on a
write action, the warning message (on linux) talked about "Error reading file".
This was because, before this fix, when opening a file we first read the header
to determine whether this is a (g)zipped file or not. If I temporarily revert
the warning fix and apply this patch, the warning message becomes (as
expected):
  warning: Couldn't write to file /home/wouter/dirasdsk/: Error opening file "/home/wouter/dirasdsk/"

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12874 7512d703-071c-0410-b5c2-d2bdd895b7d1