Fixed ARM Thumb-2 compile error
The Debian build for openmsx (0.9.0-1) on armhf failed:
  https://buildd.debian.org/status/fetch.php?pkg=openmsx&arch=armhf&ver=0.9.0-1&stamp=1348664233&file=log
It gave errors like:
  {standard input}:4944: Error: Thumb-2 MUL must not set flags -- `muls r3,lr,r3'

And indeed 'muls' (multiply instruction that sets the status flags) is not part
of the Thumb-2 instruction set, but the 'mul' instruction is (multiply that
does not change the status flags). In the full ARM instruction set both exist.
In this case we don't need the status flags, so it's OK to use 'mul', and
that's what this patch does. (Some instructions have both an S and non-S
variant in Thumb-2 and the S variant has a shorter encoding, I guess that's why
I picked the 'muls' instead of the 'mul' instruction.)

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12881 7512d703-071c-0410-b5c2-d2bdd895b7d1