Added openmsx specific bash tab-completion
After this patch you can in the bash shell type the following, and it will
work as expected:
  openm<tab> ~/msx/roms/magi<tab> -romt<tab> Mi<tab> -machi<tab> Pana<tab>G<tab>
Or more concrete, it will complete the possible openmsx options, file and
directory names and after '-machine, -ext, -romtype' it will complete 'machine,
extension, romtype' names.

See Contrib/openmsx-complete.bash for instructions on how to activate this.

It works but it's still much too slow to be useable. This will be fixed in the
next patch.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12896 7512d703-071c-0410-b5c2-d2bdd895b7d1