Speedup new -bash option, split initialization of Reactor
Before this patch when the -bash (or --help or --version) option was used, we
still fully initialized the Reactor object. This e.g. includes parsing the
softwaredb.xml and .filecache files. This made 'openmsx -bash' way too slow to
be used interactively for bash tab-completion.

This patch moves most of the initialization of Reactor from the constructor to
a new init() method. This means that parsing of the commandline options can
start much earlier and for options like -bash (and --version, --help) we can
completely skip the more expensive initialization parts of Reactor. Though this
also means that some code (e.g. the CLI parsing helper objects) now see a not
fully initialized Reactor object, so that code had to be rewritten slightly.

This patch also cleans up the command line parsing code a bit, before it had
numbered priorities, now it has named initialization phases. This code could
still be improved further, but that's something for a different patch.

CommandLineParser still contained some code to parse custom file extensions: at
one point in the past we allowed to specify filetype extensions in the settings
file (like .di2 for diskimages). Though it was broken, probably for a very long
time already: this section in the config was parsed before the settings file
was read in. Instead of trying to fix it I just removed this code, it wasn't
that useful anyway. (This code had to be changed because of the different
initialization order of Reactor, that's why this is included in this patch).

Before this patch, processing the -bash option took about 0.4s (way too slow)
on my machine. Now it only takes 0.01s (delay is not noticeable).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12897 7512d703-071c-0410-b5c2-d2bdd895b7d1