Changed return type from bool->void for CLIOption::parseOption()
The only remaining method that returned something different than 'true' was
HelpOption::parseOption() to postpone showing the help text till after the
settings file was parsed. I guess this was done to be able to show the custom
file type extensions, but the custom extensions feature was removed (because it
was already broken for a very long time). Now all methods always return 'true',
so the return value is useless and I removed it.

Also I'm not really sure why this boolean return type existed in the first
place. Postponing parsing of certain commandline options after other options
was already (more cleanly) possible using the phase/priority mechanism.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12902 7512d703-071c-0410-b5c2-d2bdd895b7d1