Further improve RomTypeInfo stuff
When you add a new romtype to openMSX, you have to:
 - <new file>:      Implement the rom type logic
 - RomTypes.hh:     Add a new enum value
 - RomFactor.cc:    Add a case to construct the new mapper
 - RomInfo.cc:      Add a new mapping between string <-> enum value
 - RomInfoTopic.cc: Add a description and blocksize for the new mapper

The last step was often forgotten because without it the code mostly works,
only the 'openmsx_info romtype <type>' info topic has missing information on
the new mapper type.

The previous patch already slightly improved this situation by enforcing that
description and blocksize were always added together (two sub steps of the last
step). This patch moves the location of the description and blocksize from
RomInfoTopic.cc to RomInfo.cc, together with the name<->enum value. So now it
should be impossible to forget to add these. And overall the code structure is
also a bit cleaner.

During this refactoring is discovered some inconsistencies in the current code:
 - There was no description/blocksize for the ROM_MEGAFLASHROMSCCPLUS type.
 - The type ROM_MSX_AUDIO still had a description, but it was already removed
   from the rest of the code.
 - There existed enum values like ROM_PAGE03 that did not have a corresponding
   mapper name. This means that in theory the code supports such (weird)
   mappers, but there's no way to actually create them (you can't name them in
   the rom database or on the commandline). Because these are weird mappers I
   didn't add a name mapping but instead I removed the enum values.
After this refactoring these type of inconsistencies should be less easy to
create.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12910 7512d703-071c-0410-b5c2-d2bdd895b7d1