Don't explicitly number the RomType enum
We used to give explicit numbers to the values in the RomType enum.
Historically this was because we used several bit masks to deduce properties
from the enum (e.g. does this mapper contain SRAM). These bitmaps are (mostly)
gone, but the explicit numbering has remained. This patch removes the explicit
numbering for all but a few mapper types where the numeric value still matters.

The motivation for this patch was that over time we've added more and more new
mapper types and the next new numeric value we would use (33) would interfere
with the mapper type 'ROM_PAGE0' (this type is still explicitly numbered).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12911 7512d703-071c-0410-b5c2-d2bdd895b7d1