Fixed colors in openMSX console background on MAC.
See this thread:
  http://www.msx.org/forum/semi-msx-talk/openmsx/improving-openmsx-console-readability

The problem was that some of the color channels of the loaded PNG image were
swapped (this bug only triggered when the image still needed to be scaled after
it was loaded).

Some more background info: revisions 12499-12504 optimized handling of PNG
images. The idea was to directly load the PNG (via libpng) in the desired
pixelformat, so that later we don't have to perform an extra conversion step.
Before this patch the code could (only) handle RGBA and BGRA pixel formats,
this patch adds ARGB and ABGR. It also re-adds the conversion in case the video
hardware still uses some more exotic pixel format than one of these 4 (it's not
straight forward to instruct libpng to directly load in still another format
than these 4).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12913 7512d703-071c-0410-b5c2-d2bdd895b7d1