Subdirectory support for DirAsDSK
This code is still experimental. I did some basic testing on it and I didn't
find any bugs. But it needs wider testing before it should be used on
production systems.

The main change is the following: Before we represented a msx directory entry
as a single number (0-111, the entry number in the root directory). Now they
are represented as a pair of [sector-number, index within that sector]. The
data members of the DirAsDSK class have a similar change: Before we had an
array with a position for each root-directory-entry. Now we have a map indexed
on the new direntry representation (sector/idx pair).

Apart from the datastructure changes of course many routines had to be extended
to also import/export directories. But overal the code changes are not that
invasive.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12915 7512d703-071c-0410-b5c2-d2bdd895b7d1