Replace std::auto_ptr with std::unique_ptr
The class std::auto_ptr is deprecated in c++11. We used this quite a lot in
openMSX, so right before this patch, compilation of openMSX with gcc produced a
whole lot of warnings. This patch fixes those warnings.

--

Some background info on auto_ptr/unique_ptr: Both these classes are
'smart-pointers' with very similar semantics. They both represent a pointer
with unique ownership. In the whole program there should be at most one
auto/unique_ptr to the same object (there can be additional raw- pointers to
that object). Because the smart pointer represents ownership, when it goes out
of scope (so when the sole owner of the object disappears), the pointed to
object has no owner anymore and it will be deleted.

Both auto_ptr and unique_ptr are also very efficient. Their internal
representation is just a single (raw) pointer. This means that they have _no_
space or time overhead compared to manual memory management using raw pointers.

Till here auto_ptr and unique_ptr are identical, the differences occur when you
start talking about transfer of ownership. Take a look at this code:

  auto_ptr<MyClass> my_ptr = ...;
  myFunction(my_ptr);

The intention is to pass ownership of the object pointed to by 'my_ptr' from
the current function to myFunction(). So after the call to myFunction()
returns, my_ptr no longer has ownership over the object. Remember that the
internal representation of an auto_ptr is nothing more than a raw pointer, so
there literally is no spare bit available to encode the property 'this auto_ptr
has ownership or not'. In auto_ptr, a non-owning pointer is represented by a
NULL pointer. In other words taking a copy of an auto_ptr objects changes the
source object into a NULL pointer. In the example above this means that if you
later add code after the call to myFunction() that still uses the my_ptr
variable (with use I mean dereference) you will get a crash. At the very least
this behavior is unexpected and I've personally hit this a few times.

So auto_ptr has weird copy semantics, but transferring ownership is a very
useful thing to do.... How does unique_ptr solve this? The most important
difference is that copying a unique_ptr is not allowed, so the example above,
with auto_ptr replaced with unique_ptr, would simply not compile. What _is_
allowed is to _move_ a unique_ptr (moving is a new concept in c++11, it's
implemented via rvalue-references, but I won't go into detail here). So if we
rewrite the above example, it now looks like this:

  unique_ptr<MyClass> my_ptr = ...;
  myFunction(move(my_ptr));

After the call, it's now again not allowed to dereference my_ptr, but at least
now it's explicit in the code that my_ptr has been moved from.

Transfer of ownership happens quite often, luckily it's not required to
everywhere add an explicit move() annotation. I'll skip over many details, but
adding move() is only required when the moved-from object can still
potentially be (mis-)used after the transfer. So annotation is not necessary
in these two important cases:

   myFunction(factorFunctionForMyClassObjects());
or
   return my_ptr;

In the first case we pass the result of a (factory) function immediately to
another function. This does move ownership, but because there's no name for the
transfered object (technically: it's an rvalue) there's no risk of (mis)using
it later. In the second example the transfered object does have a name, but
because of the return that name goes out of scope and thus there's also no risk
of misuse.

(Side-note: in case of unique_ptr copying vs moving is a matter of correctness.
There are (many more) types where moving vs copying is a matter of efficiency.
For example it's much cheaper to move a string/vector than to copy it).

--

This is a very big patch, but it's quite simple: it mechanically changes all
auto_ptr's into unique-ptr's. It also adds move() in all locations where the
compiler says it's needed (in other words where it was needed to fix a compile
error).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12924 7512d703-071c-0410-b5c2-d2bdd895b7d1