Made SDLSurfacePtr moveable (and not copyable anymore)
SDLSurfacePtr was modeled after std::auto_ptr. The previous patch changes
auto_ptr into unique_ptr. This patch changes SDLSurfacePtr so that it follows
unique_ptr rather than auto_ptr semantics. In other words: SDLSurfacePtr is now
a movable but not copyable type. As a bonus, this actually simplifies the
implementation of SDLSurfacePtr.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12925 7512d703-071c-0410-b5c2-d2bdd895b7d1