C++11 allows to use '>>' (instead of '> >') to close a nested template
Because of a parsing ambiguity in c++-98 you had to write this:
  vector<shared_ptr<MyObject> > myVector;
in c++11 you're allowed to write it also like this
  vector<shared_ptr<MyObject>> myVector;
note the space between the two '>' characters.

The latter syntax is more in line with the rest of the coding style we use in
openMSX, so I search/replaced all occurrences of '> >' with '>>'.

--

Background info: the '>>' character sequence represents the right-shift token
in c++. The new c++11 parser has to jump through hoops to get the new syntax
parsed correctly. Though because (even a c++98) c++ parser has already to jump
through much weirder hoops and because the closing template problem is
something that many c++ newbies encounter, the c++ committee decided to allow
the new syntax. More modern languages (like 'D') side step this issue by not
using '< >' to enclose template parameters (instead D uses '( )').

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12928 7512d703-071c-0410-b5c2-d2bdd895b7d1