Replace NULL with nullptr
In c/c++ the symbol NULL is conventionally used to represent the value of a
null-pointer. Unfortunately this is only a convention (some of the system
headers have '#define NULL 0') not a built-in language feature and this
sometimes causes problems. C++11 does have a built-in symbol for the
null-pointer namely 'nullptr' and that fixes most of the problem of NULL.

This patch mechanically changes all occurrences of NULL with nullptr (and
sometimes rephrases a comment).

--

Some more background info (see
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2007/n2431.pdf) for all the
details):

As already said NULL is a #define for 0. So the first problem is that if you're
not #including the right header then NULL will not compile (though a lot of
headers seem to define it, so it almost seems like NULL is always available).
Other problems come from the fact that the compiler itself doesn't see the NULL
symbol (the preprocessor has already replaced it with 0). So for example in
this code:

  void f(int a);
  void f(char* s);
  ...
  f(NULL);

The intention of the programmer was likely to call the 2nd overload of the
function f(), but the c++ language rules say that the first version is a better
match because the _int_ literal '0' was passed.

The paper mentioned above has more examples where NULL behaves strange. In the
context of openMSX we didn't really encounter many NULL-problems, but because
c++11 has a nice solution, I decided to use it anyway.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12929 7512d703-071c-0410-b5c2-d2bdd895b7d1