Compile with -Wzero-as-null-pointer-constant and fix all warnings
This is a followup on revision 12929. Gcc-4.7.x has a new option to generate
warnings when an int literal '0' is passed to a function that expects a
pointer. C++11 has a name for the null-pointer (namely nullptr) so it's much
clearer to always explicitly say 0 for integers and nullptr for pointers.

I've enabled this warning by default (only on compilers that do support it).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12931 7512d703-071c-0410-b5c2-d2bdd895b7d1