Replace our own shared_ptr class with std::shared_ptr
We used to have our own implementation of shared_ptr. C++11 has a standard
implementation of this class, and it's much better optimized and offers many
more features. The part of shared_ptr that was implemented in our version was
fully compatible with the standard. So switching was a simple matter of
changing the #include and prepending 'std::' (or add the appropriate using
directive).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12937 7512d703-071c-0410-b5c2-d2bdd895b7d1