Use make_shared() to create objects hold by shared_ptr
One of the optimizations that std::shared_ptr offers over our (already removed)
shared_ptr class is 'std::make_shared()'. When you new an object and
immediately pass the raw pointer to a shared_ptr (which is a very good idea),
it's more efficient to use make_shared(). So replace
  shared_ptr<MyClass>(new MyClass(i, 3));
with
  make_shared<MyClass>(i, 3);

--

Background info: shared-ptr is a reference-counted smart pointer. There can be
many pointers to the same object, all these pointers share ownership of this
object. Only when the last shared_ptr to an object goes away, that object will
be deleted. In contrast to unique_ptr (or auto-ptr) a shared_ptr does have
space and time overhead compared to a raw-pointer.

All shared_ptr's to the same (shared) object must also share the reference
counter. This counter cannot be stored in the shared_ptr object itself, it also
cannot be stored in the pointed-to-object. Instead this counter is usually stored
in a separate heap-allocated object. The shared_ptr objects themselves usually
have both a pointer to the pointed-to object and to this shared counter (there
are other possibilities, but this approach is common).

Now the make_shared-optimization: make_shared() has to both allocate/construct
the pointed-to object and create a shared_ptr to this object (at this point
it's still the only pointer to this object, so the shared counter also has to
be allocated). The make_shared() function can allocate both memory chunks using
a single heap-request. This has less memory management overhead. This also
ensures that the object and the counter are next to each other in memory which
is good for cache locality. This brings the performance of std::shared_ptr
almost to the same level as an intrusive reference counted pointer (with the
counter stored in the pointed to object).

--

This patch changes many 'shared_ptr(new ...)' into make_shared patterns. It
also changes the EventDistributor::distributeEvent() method to take a
shared_ptr instead of a raw pointer.

This patch also (in a very small way) makes use of the new c++11 'auto'
feature, compare:
   shared_ptr<MyClass> my_ptr = make_shared<MyClass>(...);
with
   auto my_ptr = make_shared<MyClass>(...);
Expect to see many more auto-related changes in future patches.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12938 7512d703-071c-0410-b5c2-d2bdd895b7d1