Replaced our STATIC_ASSERT macro with the c++11 static_assert feature
Our STATIC_ASSERT macro was a bit of a kludge, we already had to change it a
couple of times to keep it working (warning free) with newer compiler versions.
Luckily c++11 offers a standard static_assert feature. Unfortunately our
STATIC_ASSERT assert macro was not fully compatible with c++11's static_assert.
The latter requires a 2nd string-literal that will be printed as compiler error
message when the assert triggers. So this patch adds these messages everywhere.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12941 7512d703-071c-0410-b5c2-d2bdd895b7d1