Also initialize member variables using make_unique()
Replace in the initializer list the pattern
  memberVar(new MyClass(...))
with
  memberVar(make_unique<MyClass>(...))

The generated code should be identical in both cases. The advantage is that
this makes the kind of memory management explicit and that it gets rid of the
low level 'new' keyword (the goal is to have very few 'new'/'delete' keywords
left in openmsx).

Disadvantage is that this change makes the lines slightly longer (the previous
make_unique() changes often made lines shorter). This pushed some lines over
our (soft) 80-char-per-line limit, though many lines already were longer to
start with. This patch reformats the code, so if you look at the number of
added/removed lines in this patch, it seems like this patch adds a lot of code,
but that's only because of reformatting.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12953 7512d703-071c-0410-b5c2-d2bdd895b7d1