Replace vector<Foo*> with vector<unique_ptr<Foo>>
Often containers logically own the objects they contain, sometimes those
objects are stored by raw-pointer. This patch changes that to unique_ptr. This
makes the ownership explicit. It also allows to remove the manual
deletion(-loop), or it fixes a leak when the cleanup code was missing. This is
also safer in case of exceptions.

--

Some background info: in c++98 code containers of owning raw-pointers occur
fairly often. In c++98 the class std::auto_ptr (now deprecated) was used to
manage ownership. Though a big disadvantage of auto_ptr was that it could not
be stored in STL containers. The reason is that auto_ptr has a weird
copy-constructor. When you copy an auto_ptr you destroy the original. STL
containers (certainly in C++98) often copy their elements (e.g. when a vector
has run out of reserved space it has to allocate a new memory block and copy
the elements from the old to the new memory block). When an STL container
would start copying auto_ptr objects weird stuff would happen, so luckily
such code doesn't compile.

This problem is fixed with the c++11 unique_ptr class. A unique_ptr still
cannot be copied, but it can be moved. The STL containers in C++11 are updated
to (where possible) move their elements instead of copy them. (Very briefly and
very imprecise: a move is a bit like a copy where you don't care about the
original, usually because you know the original will be destroyed soon. A move
can (for some types) be cheaper than a copy because it can 'steal' the internal
guts from the source object)

--

In openMSX containers that own their elements were either implemented:
 - with raw-pointers plus manual memory management
 - with shared_ptr's (gives cleaner code, but it has some overhead)
This patch changes the code to use unique_ptr as much as possible. This gives
almost as clean code as in the shared_ptr case, but it doesn't add any overhead
compared to the raw-pointer case.

There are 3 other things that happen in this patch that are worth mentioning.
(These changes were needed for the unique_ptr refactoring, so I couldn't
easily split them to separate patches).

1) Use of the 'auto' keyword
I already used this in previous patches, but I didn't really explain it yet.
(Roughly speaking) 'auto' can be used in all locations where the c++ language
syntactically expects a type, and the compiler already knows what that type
should be. An example:
  vector<int>::iterator it = find(myVec.begin(), myVec.end(), 42);
this can be simplified to:
  auto it = find(myVec.begin(), myVec.end(), 42);
According to the c++ syntax, you need to specify a type for the variable 'it'.
Though if you type anything other than 'vector<int>::iterator', the compiler
will complain and tell you you should have typed 'vector<int>::iterator'. These
kind of compiler messages can be frustrating ;-) Now there's a way to tell the
compiler: "Hey, you already know the type, just use that one".

So this feature doesn't bring anything new, but it can drastically simplify the
code. Or it can make the code more future proof. For example:
   auto str = object.getString();
   doSomethingWithString(str);
I don't really care whether getString() returns a std::string or a string_ref
or still some other string like object. I only care that I can pass this object
to doSomethingWithString() (which obviously has to accept string-like objects).
So by using 'auto', it's possible to change the return type of getString()
without having to necessarily adjust all callers of this method.

There are still _many_ locations in openMSX where we could simplify the code
by using 'auto'.

2) Return a collection by value
C++98 code often contained code like
  vector<string> result;
  getCollection(result);
instead of the more natural
  vector<string> result = getCollection();
The reason is efficiency: in C++98 the 2nd form required making (one or more!)
copies of the container. This overhead can be significant for large containers.
In c++11 STL collections can be moved instead of copied and all this overhead
is gone.

This patch only uses this more natural syntax for a single method (because that
method had to be changed for the unique_ptr refactoring):
	XMLElement::getChildren(string_ref)
expect more cleanups like this in future patches.

3) Use of lambda expressions
Lambdas are a big new feature in c++11, they don't bring anything fundamentally
new to c++, but they do offer a much more convenient syntax compared to c++98.
I'll only give an example, if you want you can google the full details of c++
lambda expressions. But if you know lambdas from other languages it should be
familiar

C++98 code with explicit loop:
	Person* findPersonByName(const string& name) {
		for (vector<Person*>::iterator it = persons.begin();
		     it != persons.end(); ++it) {
			if ((*it)->getName() == name) {
				return *it;
			}
		}
		return NULL;
	}

STL has a nice set of algorithms. The 'find_if' algorithm would be a perfect
fit here. Though unfortunately using it is not so simple in c++98:

C++98 code using the find_if algorithm:
	struct EqualName {
		EqualName(const string& name_) : name(name_) {}
		bool operator(const Person* p) const {
			return p->getName() == name;
		}
		const string name;
	};
	...
	Person* findPersonByName(const string& name) {
		vector<Person*>::iterator it = find_if(
			persons.begin(), persons.end(), EqualName(name));
		return (it != persons.end()) ? *it : NULL;
	}

So we've used the STL algorithm, but in order to do so we had to write a whole
lot of boiler-plate code: we had to define a functor (the EqualName struct) and
even worse that struct had to be defined outside of the findPersonByName()
function. Thus the logic of this functions is spread over different locations in
the source code.

C++11 code with find_if and lambda function:
	Person* findPersonByName(const string& name) {
		auto it = find_if(persons.begin(), persons.end(),
			[&](Person* p) { return p->getName() == name; });
		return (it != persons.end()) ? *it : NULL;
	}

So again this doesn't bring anything new to c++ (the 2nd and 3rd variant really
do exactly the same), it's just a much more convenient way to write functors.
Notice how the body of the lambda expression and the body of the functor are
the same. Also notice how the lambda expression 'captures' the 'name' parameter
(in the functor example this capturing had to be coded explicitly).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12955 7512d703-071c-0410-b5c2-d2bdd895b7d1