Fixed(?) visual studio specific compiler errors
There were two types of errors:

1) find_if -> std::find_if
I think the compiler should have found the find_if function (located in the
std namespace) even without it being explicitly qualified. Because of argument
dependent lookup (ADT). At least in gcc it works like this.

2) missing (auto-generated?) move-constructor and move assignment
We have a struct with some member variables and one of these vars is a
unique_ptr. A unique_ptr is not copyable/assignable (but it is movable).
Because of this the whole struct is also not copyable. However the struct can
be made movable, I'm not sure but *I*think* in this simple case the compiler
should have auto-generated the appropriate move-constructor and move-assignment
operator. At least gcc seems to generate them, but visual studio apparently
does not. So I manually added them.

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12958 7512d703-071c-0410-b5c2-d2bdd895b7d1