Use return value instead of output-parameter
As already briefly explained in revision 12955, in c++11 there's no
(performance) reason anymore to return a collection from a function via an
output parameter instead of via the more natural return value. This allows for
some nice cleanups in the code.

In general passing stuff by value is much cheaper in c++11 and in some cases
even faster than the old return-via-out-parameter idiom. See this blog for
details:
  http://cpp-next.com/archive/2009/08/want-speed-pass-by-value/

--

There's one small extra c++11 feature that this patch uses that I haven't
explained yet in earlier commits: it uses the new std::begin() and std::end()
functions. Very briefly:
   vector<int>::iterator it = myvector.begin();
can now also be written like:
   vector<int>::iterator it = begin(myvector);
So instead of using the member function begin() (and end()) there's now also a
non-member function begin() (and end()) that does exactly the same thing. So
what's the advantage (beside it being one char shorter)? The advantage of the
non-member functions is that they can also be implemented for plain arrays
(that's what this patch uses them for), or even for non STL-like custom
containers (it's always possible to add global overloads for begin()/end(), but
it may not be possible to extend the custom container class with begin()/end()
member functions).

Some people suggest to replace all uses of begin()/end() member functions with
the non-member variants. Though at the moment I don't yet see the advantage of
doing such a global search/replace. Maybe in the future when we have more
experience with c++11 (and c++11 coding style).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@12961 7512d703-071c-0410-b5c2-d2bdd895b7d1